Remove Adding CNB docs
We decided to remove the optionality of this feature and opt all customers in.
Therefore, we've removed the radio button, so the installation docs are no longer valid.

I decided to move some of the other information about CNBs -- such as the overview, the removal process,
and the known limitations, to the CNB overview.